+++ forui/lib/src/widgets/select_tile_group/select_tile_group.dart
@@ -1,6 +1,7 @@
 // ignore_for_file: invalid_use_of_protected_member
 
 import 'package:flutter/foundation.dart';
+import 'package:flutter/gestures.dart';
 import 'package:flutter/widgets.dart';
 
 import 'package:forui/forui.dart';
@@ -17,6 +18,8 @@ import 'package:forui/src/widgets/select_tile_group/select_tile.dart';
 /// * [FSelectTile] for a single select tile.
 /// * [FTileGroupStyle] for customizing a select group's appearance.
 class FSelectTileGroup<T> extends FormField<Set<T>> with FTileGroupMixin<FTileMixin> {
+  static bool _predicate(BuildContext _, int __) => true;
+
   static Widget _errorBuilder(BuildContext context, String error) => Text(error);
 
   /// The controller.
@@ -24,7 +27,35 @@ class FSelectTileGroup<T> extends FormField<Set<T>> with FTileGroupMixin<FTileMi
   /// See:
   /// * [FRadioSelectGroupController] for a single radio button like selection.
   /// * [FMultiSelectGroupController] for multiple selections.
-  final FSelectGroupController<T> controller;
+  final FSelectGroupController<T> groupController;
+
+  /// The scroll controller used to control the position to which this group is scrolled.
+  ///
+  /// Scrolling past the end of the group using the controller will result in undefined behaviour.
+  ///
+  /// It is ignored if the group is part of a merged [FTileGroup].
+  final ScrollController? scrollController;
+
+  /// The cache extent in logical pixels.
+  ///
+  /// Items that fall in this cache area are laid out even though they are not (yet) visible on screen. It describes
+  /// how many pixels the cache area extends before the leading edge and after the trailing edge of the viewport.
+  ///
+  /// It is ignored if the group is part of a merged [FTileGroup].
+  final double? cacheExtent;
+
+  /// The max height, in logical pixels. Defaults to infinity.
+  ///
+  /// It is ignored if the group is part of a merged [FTileGroup].
+  ///
+  /// ## Contract
+  /// Throws [AssertionError] if [maxHeight] is not positive.
+  final double maxHeight;
+
+  /// Determines the way that drag start behavior is handled. Defaults to [DragStartBehavior.start].
+  ///
+  /// It is ignored if the group is part of a merged [FTileGroup].
+  final DragStartBehavior dragStartBehavior;
 
   /// The style. Defaults to [FThemeData.tileGroupStyle].
   final FTileGroupStyle? style;
@@ -50,14 +81,15 @@ class FSelectTileGroup<T> extends FormField<Set<T>> with FTileGroupMixin<FTileMi
   /// The semantic label used by accessibility frameworks.
   final String? semanticLabel;
 
-  /// The items.
-  final List<FSelectTile<T>> children;
-
   /// Creates a [FSelectTileGroup].
   FSelectTileGroup({
-    required this.controller,
-    required this.children,
+    required this.groupController,
+    required List<FSelectTile<T>> children,
+    this.scrollController,
     this.style,
+    this.cacheExtent,
+    this.maxHeight = double.infinity,
+    this.dragStartBehavior = DragStartBehavior.start,
     this.divider = FTileDivider.indented,
     this.label,
     this.description,
@@ -82,7 +114,11 @@ class FSelectTileGroup<T> extends FormField<Set<T>> with FTileGroupMixin<FTileMi
             };
 
             return FTileGroup(
+              controller: scrollController,
               style: groupStyle,
+              cacheExtent: cacheExtent,
+              maxHeight: maxHeight,
+              dragStartBehavior: dragStartBehavior,
               divider: divider,
               label: label,
               enabled: enabled,
@@ -92,11 +128,86 @@ class FSelectTileGroup<T> extends FormField<Set<T>> with FTileGroupMixin<FTileMi
               children: [
                 for (final child in children)
                   FSelectTileData<T>(
-                    controller: controller,
-                    selected: controller.contains(child.value),
+                    controller: groupController,
+                    selected: groupController.contains(child.value),
                     child: child,
                   ),
               ],
+            );
+          },
+        );
+
+  /// Creates a [FSelectTileGroup] that lazily builds its children.
+  ///
+  /// The [tileBuilder] is called for each tile that should be built. It will be called only for indices <= [count] if
+  /// [count] is specified.
+  ///
+  /// The [predicate] returns true if the tile at the given index should be built. It may be called more than once for
+  /// the same index.
+  ///
+  /// The [count] is the number of tiles to build. If null, [tileBuilder] will be called until [predicate] returns false.
+  ///
+  /// ## Notes
+  /// May result in an infinite loop or run out of memory if:
+  /// * [count] is null and [tileBuilder] always provides a zero-size widget, i.e. SizedBox(). If possible, provide
+  ///   tiles with non-zero size, return null from builder, or set [count] to non-null.
+  ///
+  /// * placed in a parent widget that does not constrain its size, i.e. [Column].
+  FSelectTileGroup.builder({
+    required this.groupController,
+    required FSelectTile<T> Function(BuildContext, int) tileBuilder,
+    bool Function(BuildContext context, int index) predicate = _predicate,
+    int? count,
+    this.scrollController,
+    this.style,
+    this.cacheExtent,
+    this.maxHeight = double.infinity,
+    this.dragStartBehavior = DragStartBehavior.start,
+    this.divider = FTileDivider.indented,
+    this.label,
+    this.description,
+    this.errorBuilder = _errorBuilder,
+    this.semanticLabel,
+    super.onSaved,
+    super.validator,
+    super.initialValue,
+    super.forceErrorText,
+    super.enabled = true,
+    super.autovalidateMode,
+    super.restorationId,
+    super.key,
+  }) : super(
+          builder: (field) {
+            final state = field as _State;
+            final groupStyle = style ?? state.context.theme.tileGroupStyle;
+            final (labelState, error) = switch (state.errorText) {
+              _ when !enabled => (FLabelState.disabled, null),
+              final text? => (FLabelState.error, errorBuilder(state.context, text)),
+              null => (FLabelState.enabled, null),
+            };
+
+            return FTileGroup.builder(
+              controller: scrollController,
+              style: groupStyle,
+              cacheExtent: cacheExtent,
+              maxHeight: maxHeight,
+              dragStartBehavior: dragStartBehavior,
+              count: count,
+              divider: divider,
+              label: label,
+              enabled: enabled,
+              description: description,
+              error: error,
+              semanticLabel: semanticLabel,
+              predicate: predicate,
+              tileBuilder: (context, index) {
+                final child = tileBuilder(context, index);
+                return FSelectTileData<T>(
+                  controller: groupController,
+                  selected: groupController.contains(child.value),
+                  child: child,
+                );
+              },
             );
           },
         );
@@ -108,8 +219,12 @@ class FSelectTileGroup<T> extends FormField<Set<T>> with FTileGroupMixin<FTileMi
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
     properties
-      ..add(DiagnosticsProperty('controller', controller))
+      ..add(DiagnosticsProperty('groupController', groupController))
+      ..add(DiagnosticsProperty('scrollController', scrollController))
       ..add(DiagnosticsProperty('style', style))
+      ..add(DoubleProperty('cacheExtent', cacheExtent))
+      ..add(DoubleProperty('maxHeight', maxHeight))
+      ..add(EnumProperty('dragStartBehavior', dragStartBehavior))
       ..add(DiagnosticsProperty('divider', divider))
       ..add(ObjectFlagProperty.has('errorBuilder', errorBuilder))
       ..add(StringProperty('semanticLabel', semanticLabel));
@@ -120,38 +235,38 @@ class _State<T> extends FormFieldState<Set<T>> {
   @override
   void initState() {
     super.initState();
-    widget.controller.addListener(_handleControllerChanged);
+    widget.groupController.addListener(_handleControllerChanged);
   }
 
   @override
   void didUpdateWidget(covariant FSelectTileGroup<T> old) {
     super.didUpdateWidget(old);
-    if (widget.controller == old.controller) {
+    if (widget.groupController == old.groupController) {
       return;
     }
 
-    widget.controller.addListener(_handleControllerChanged);
-    old.controller.removeListener(_handleControllerChanged);
+    widget.groupController.addListener(_handleControllerChanged);
+    old.groupController.removeListener(_handleControllerChanged);
   }
 
   @override
   void didChange(Set<T>? values) {
     super.didChange(values);
-    if (!setEquals(widget.controller.values, values)) {
-      widget.controller.values = values ?? {};
+    if (!setEquals(widget.groupController.values, values)) {
+      widget.groupController.values = values ?? {};
     }
   }
 
   @override
   void reset() {
     // Set the controller value before calling super.reset() to let _handleControllerChanged suppress the change.
-    widget.controller.values = widget.initialValue ?? {};
+    widget.groupController.values = widget.initialValue ?? {};
     super.reset();
   }
 
   @override
   void dispose() {
-    widget.controller.removeListener(_handleControllerChanged);
+    widget.groupController.removeListener(_handleControllerChanged);
     super.dispose();
   }
 
@@ -161,8 +276,8 @@ class _State<T> extends FormFieldState<Set<T>> {
     // In the case where a controller has been passed in to this widget, we register this change listener. In these
     // cases, we'll also receive change notifications for changes originating from within this class -- for example, the
     // reset() method. In such cases, the FormField value will already have been set.
-    if (widget.controller.values != value) {
-      didChange(widget.controller.values);
+    if (widget.groupController.values != value) {
+      didChange(widget.groupController.values);
     }
   }